requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
httpx
pymongo
pandas>=2.0
numpy>=1.24
sqlalchemy>=2.0
pymysql>=1.1
scikit-learn>=1.3
xgboost>=2.0
joblib>=1.3
python-dotenv>=1.0